Management functionality (#10)
* added show parking statuses live for manager (updates every 5 minutes)

* improve styling and added booking id

* added muli building option

* parking statuses display based of building id

* manager get builing_id = 0

* parking management (add,update,remove) - first step

* complete parking management by manager

+++ backend/db/db_types/building_table_types.py
@@ -1,5 +1,6 @@
-from sqlalchemy import Column, Integer, String, Float # type: ignore
-from sqlalchemy.ext.declarative import declarative_base # type: ignore
+from sqlalchemy import Column, Integer, String, DECIMAL, DateTime, Boolean
+from sqlalchemy.ext.declarative import declarative_base
+
 
 Base = declarative_base()
 
@@ -9,7 +10,7 @@ class Building(Base):
     building_id = Column(Integer, primary_key=True, autoincrement=True)
     building_name = Column(String, nullable=False)
     building_address = Column(String, nullable=False)
-    building_address_latitude = Column(Float, nullable=False)
-    building_address_longitude = Column(Float, nullable=False)
+    building_address_latitude = Column(DECIMAL(9,6), nullable=False)
+    building_address_longitude = Column(DECIMAL(9,6), nullable=False)
     manager_id = Column(Integer, nullable=False)
